Fix PR per-check issue
